Rm EventLoopManager since it's not needed anymore
